Add files via upload
Model tunning 

There is no one solution or approach that fits all. There

are several factors over and above the model performance that can impact the decision to choose a machine learning algorithm.

I created a syntetic dataset to see how is the best approach for this dataset using linear and polynomial regression. As You will see, this is a alway to automatize some models selections. I used to this for every dataset I was using. Now I can improve my research modeling some algorithms and just changing the input of variables. I hope this can help many researchs.